created File Manager auto load class.
refactored directory handling such as open_project from graph_editor to
file_manager. This is working, plan on doing the same for project
creation.

refactored checking for save request to the file_manager and removed it
from the level_container. Added a file type enum to the file manager,
and upon save request, checks the type of the currently focused control
node, and emits the active type trying to be saved.

created various static helper methods in file_manager for opening and
checking if a file exists.

renamed the save_level action in input mapping to be more agnostic and
general purpose.

The save window context aware of the type of file it is saving.