#SCRUM-45 #Review Fixed a few bugs in this commit, when changing levels, the level wasn't resetting it's selected node, but node details was clearing it's active node on level change, thus resulting in the crash for checking if a property name already exist's on nil node.
Both property editor and node details are cleared of any work done on the previously active level.
level.gd : changes consisted of not emitting a single in _ready for the name change, i believe it was causing interment errors., and just setting the level name to it's default name in _ready. the load_level function call's set_tab_name when setting the name of the loading level.

levels_container.gd: changes consisted of removing an emitting signal from _ready, as it was causing crashes in other nodes that weren't in the scene tree, just yet.

fixed a typo in file_manager.gd