Setting the types for nodes. Implemented slot color and active slots for the various node types, and only allowing the correct connections. Connections are: Entry -> Transit, Exit Transit -> Transit, Exit Link -> Entry, Transit
Setting the selected type from the new node window when the node is created. Set the slots by their type on
load, and when they are added to the level. This step had to be completed in order to move to the storyline
tree feature. I also had an issue with level's previously saved before I started tracking the node id's that I
had to fix. However, that wouldn't fix the node's that were already in the level.