Command Refactor
Saving levels as JSON instead of binary.
Created global FileTypes enum.
Fixed some broken file paths where I was using const string literals before learning about export_file.
Added create new file scene.
Added file path box to save window. However, I'm not using this and will be removed soon. Moved to forcing a file path and name on creation.
The set_command method on CommandInvoker return's itself so method calls to execute_command can be chained.
added new level command, as this is handled in multiple places and it is working. Still needs some polish.
added open file command, this is working.
save command is working as intended. Added a callback signal so the invoking object will know when the save operation is completed, every context object should have a get_resource_path method in order to be saved.
changed the level thumbnail to use the open file command when double clicked.
thumbnail_base holds a reference to the command_invoker class so any thumbnail will have access to executing commands.
changed the method name from save_level to save. This way, objects that need to be serialized will have a general, agnostic name for saving.
The menu_bar is using the new level command for creating new level's, and will use the open command as well after more clean up.
hiding the root object in the story lines inspector. Also, connected to a new signal on_level_created from the levels_container.
Wrote a function to parse the vector 2 position for the node from a string to vector2 when loading the node data back in.
FileIO.gd now opens files, distinguishing file type from its extension, and calls a relevant function. Although it will experience some cleanup upon the next commit, as the save window is irrelevant. This was a huge change, and a lot of code is now unused and will slowly be removed. New features are on pause as I clean up this scalability issue. I have not experienced any bugs at the time of this commit; functionality is back to where it was pre-command pattern refactor.