Updating Level Ids, saving and loading level id.
const variables are implied to be static in GDScript, thus updated export_command to use the const values instead of the static getter functions.

graph_editor.gd tracks the current level ID and increments it, writing it to the project file every time a level is created.

Levels set their ID when they are created and added to the tree.

updated the way data is being saved; it is now formatted for easier readability. The story project file now uses JSON format as well.